+++ pre_commit_pep639/core_utils.py
@@ -1,4 +1,6 @@
+import tomllib
 import typing as t
+from pathlib import Path
 
 
 def extract_metadata_base(toml_contents: dict[str, t.Any]) -> dict[str, t.Any] | None:
@@ -20,3 +22,21 @@ def extract_metadata_base(toml_contents: dict[str, t.Any]) -> dict[str, t.Any] |
     else:
         # Project metadata could not be found
         return None
+
+
+def get_license_globs(toml_file: Path) -> list[str]:
+    """
+    Check the provided TOML file its `license-files` field.
+
+    The provided TOML file is assumed to contain a `project` table, which contains a list of glob
+    patterns in a `license-files` field.
+    """
+    with toml_file.open("rb") as f:
+        contents = tomllib.load(f)
+
+    base = extract_metadata_base(contents)
+    if base is None:
+        # Unsupported metadata spec
+        return []
+
+    return base.get("license-files", [])  # type: ignore[no-any-return]