+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 Versions follow [Semantic Versioning](https://semver.org/spec/v2.0.0.html) (`<major>`.`<minor>`.`<patch>`)
 
 ## [v0.2.0]
+### Added
+* #2 Add `check-glob-syntax` hook
+
 ### Changed
 * #2 Rename `check-license-glob` hook to `check-glob-resolve`